refactor(category): create category dir for category module
