+++ src/modules/car/useCases/listAvailableCars/ListAvailableCarsUseCase.spec.ts
@@ -1,7 +1,7 @@
 import 'reflect-metadata';
 
 import { FakeCarRepository } from '@modules/car/repositories/fakes/FakeCarRepository';
-import { FakeCategoryRepository } from '@modules/car/repositories/fakes/FakeCategoryRepository';
+import { FakeCategoryRepository } from '@modules/category/repositories/fakes/FakeCategoryRepository';
 import { ListAvailableCarsUseCase } from './ListAvailableCarsUseCase';
 
 describe('ListCarUseCase', () => {
