+++ src/modules/category/repositories/ICategoryRepository.ts
